projects/BankIDAuth/code/bankid-auth/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
pybankid==0.12.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
